requirements copy.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
httpx==0.28.1
pydantic==2.11.3
python-multipart==0.0.20
motor==3.6.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
